fix riverpod 3.0 keepalive lint
we got this warning:
If a provider is declared as `keepAlive`, it can only use providers that are also declared as `keepAlive. Either stop marking this provider as `keepAlive` or remove `keepAlive` from the used provider.

we don't need to keepalive these providers anymore i think. seems
everything still works fine